# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's JWT/Swagger",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://api-jwt.test"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "endpoint para autentificar",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "422": {
                        "description": "Datos incorrectos."
                    },
                    "default": {
                        "description": "unauthorized."
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "registra un nuevo usuario",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "unauthorized.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Cierra la sesión del usuario logueado.",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "unauthorized.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Renueva el token del usuario logueado.",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::refresh",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "unauthorized.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user-profile": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Obtiene los datos del usaurio logueado.",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::userProfile",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "unauthorized.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Obtiene el listado de usuarios registrados.",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::index",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "unauthorized.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}